==> 1_extraccion.ipynb <==
# Librerías y configuraciones


import gzip
import ast
import json
import pandas as pd
import warnings

pd.set_option('display.max_colwidth', 100)
warnings.filterwarnings('ignore')

---

# Funciones auxiliares.

def duplicados_de(df):
    df_columns = df.columns.tolist()
    for i in df.columns:
        print( 'Duplicados de ',i,': ', df.duplicated(subset=i,keep=False).sum())

----------------

-----------

# Extracción de datos

----

# user_reviews

----------------

Carga de datos

datos = []

# Usamos with porque es una declaración de contexto en Python. 
# Se utiliza para asegurar que un recurso se abra y se cierre correctamente.
# Usamos gzip para abrir y descomprimir el archivo
with gzip.open("data/raws/user_reviews.json.gz", 'rt', encoding='utf-8') as archivo_descomprimido:
    
    for linea in archivo_descomprimido:
        # eval() evalua una cadena de texto (string) que contiene una expresión o código Python.
        # En este caso evalua los strings de cada línea y nos devuelve un diccionario por línea
        dato = eval(linea)
        # Los agregamos a la lista 'datos'
        datos.append(dato)

user_reviews_raw = pd.DataFrame(datos)
user_reviews_raw.head(2)

----------

Análisis de duplicados y nulos.

user_reviews_raw.info()

La función .info() nos dice que todos los valores de la tabla son no nulos.

Trabajaremos en una tabla user_reviews para no afectar los datos crudos.

user_reviews = user_reviews_raw

user_reviews['reviews'] = user_reviews['reviews'].apply(lambda x: json.dumps(x)) 
# convertimos los datos de reviews en str para poder usar duplicated()
print('Filas duplicadas: ',user_reviews.duplicated().sum())

Tenemos 313 filas duplicadas que eliminaremos.

user_reviews = user_reviews.drop_duplicates().reset_index(drop=True)

Ahora analizaremos las filas que poseen campos duplicados por columna.

duplicados_de(user_reviews)

(
    user_reviews[user_reviews.duplicated(subset='user_id',keep=False)]
)

Podemos observar que este usuario posee un registro vacio, lo que nos permitió notar que tenemos reviews en blanco marcados por '[]'. Procederemos a eliminarlos.

user_reviews = user_reviews[user_reviews.reviews !='[]'].reset_index(drop=True)

print( 'Duplicados de user_id: ', user_reviews.duplicated(subset='user_id',keep=False).sum())
print( 'Duplicados de user_url: ', user_reviews.duplicated(subset='user_url',keep=False).sum())
print( 'Duplicados de reviews: ', user_reviews.duplicated(subset='reviews',keep=False).sum())


La tabla ya no posee datos nulos ni duplicados.

# user_reviews

---------

----

# users_items

----

# Abre el archivo .gz y lo descomprime
datos = []
with gzip.open('data/raws/users_items.json.gz', 'rt', encoding='utf-8') as archivo_descomprimido:
    for linea in archivo_descomprimido:
        dato = eval(linea)
        datos.append(dato)

users_items_raw = pd.DataFrame(datos)

users_items_raw.head(2)

users_items_raw.info()

Podemos ver que todos los datos de user_items son no nulos.

----

Análisis de duplicados y nulos.

users_items = users_items_raw
users_items['items'] = users_items['items'].apply(lambda x: json.dumps(x)) # convertimos los datos de reviews en str para poder usar duplicated()

# Revisamos los datos de la columna items
users_items_raw.at[0, 'items'][0:2]

print('Filas duplicadas: ',users_items.duplicated().sum())

Tenemos 657 filas duplicadas que eliminaremos.

users_items = users_items.drop_duplicates().reset_index(drop=True)

duplicados_de(users_items)

Podemos observar que tenemos 53 user_id duplicados y 53 user_url duplicados.

users_items[users_items.duplicated(subset='user_id',keep=False)].sort_values('user_id').head(5)

cadena1 = str(users_items.at[4075,'items'])
cadena2 = str(users_items.at[4075,'items'])
i=0
for caracter in cadena1:
    
    if cadena1[i] != cadena2[i]:
        print (i)
    i+=1

cadena1 = (users_items.at[4075,'items'])
cadena2 = (users_items.at[14958,'items'])
rango = min(len(cadena1), len(cadena2))

# Itera sobre las posiciones hasta la longitud mínima
for i in range(rango):
    if cadena1[i] != cadena2[i]:
        print(f'Las cadenas difieren en el carácter {i}: {cadena1[i-20:i+30]} != {cadena2[i-20:i+30]}')
        break

Podemos ver que estas filas difieren en la cantidad de horas que un usuario consumió un item. Luego de desanidar esta columna se decidirá que hacer con los datos duplicados en el campo user_id.

Exploraremos si existen listas vacias en los datos de la columna items.

users_items[users_items['items'] == '[]']

Podemos observar que tenemos usuarios que no han consumido ningun item por lo que items_count es 0 e items posee '[]'. Eliminaremos estos usuarios del df principal pero guardaremos esta información en una tabla.

users_0items_count = users_items[users_items['items'] == '[]']

users_items = users_items[users_items['items'] !='[]'].reset_index(drop=True)

revisemos nuevamente los duplicados

duplicados_de(users_items)

Tenemos 53 duplicados de los que nos ocuparemos en un análisis posterior. Los valores duplicados de items_count e items son esperados así que los mantendremos.

---------

-----

# steam_games

Cargamos los datos del archivo steam_games.json.gz

datos = []

# Usamos with porque es una declaración de contexto en Python. 
# Se utiliza para asegurar que un recurso se abra y se cierre correctamente.
# Usamos gzip para abrir y descomprimir el archivo games_ruta
with gzip.open("data/raws/steam_games.json.gz", 'rt', encoding='utf-8') as archivo_descomprimido:
    for linea in archivo_descomprimido:
        # Cargamos cada línea como un objeto JSON
        obj_json = json.loads(linea)
        # Los agregamos a la lista 'datos'
        datos.append(obj_json)
# creamos un df a partir de la lista 'datos'
gamesRaw = pd.DataFrame(datos)

Exploramos esos datos

gamesRaw.columns

gamesRaw.info()

-----

Vamos a revisar la cantidad de datos nulos

(
    gamesRaw
    .isna()
    .sum()
)

Podemos notar que las las columnas 'user_id', 'steam_id','items' e 'items_count' poseen la misma cantidad de datos nulos.

veamos la cantidad de filas que poseen nulos en todas sus columnas

cant_filas_solo_nulos = gamesRaw[gamesRaw.isna().all(axis=1)].shape[0]
print (cant_filas_solo_nulos, 'filas con todos sus valores nulos')

Revisaremos ahora las filas que no tienen valores nulos en user_id

gamesRaw[gamesRaw['user_id'].notna()].info()

Podemos ver que todas las filas no contienen valores no nulos en user_id, poseen valores nulos en las primeras 15 columnas de la tabla. Por lo que entendemos que estamos en presencia de dos tablas independientes unidas.

-------

Revisaremos si la tabla items obtenida del archivo 'users_items.json.gz' es la que se encuentrea anexada a stam_games

gamesRaw.columns

users_items.columns

items_anexados_a_steam_games = gamesRaw[['user_id','items_count','steam_id','items']]
items_anexados_a_steam_games = items_anexados_a_steam_games.dropna()
print('users_items_raw = items_anexados_a_steam_games?', users_items_raw.equals(items_anexados_a_steam_games))
print('users_items_raw.user_id = items_anexados_a_steam_games.user_id?', users_items_raw.user_id.equals(items_anexados_a_steam_games.user_id))
print('users_items_raw.items_count = items_anexados_a_steam_games.items_count?', users_items_raw.items_count.equals(items_anexados_a_steam_games.items_count))
print('users_items_raw.steam_id = items_anexados_a_steam_games.steam_id?', users_items_raw.steam_id.equals(items_anexados_a_steam_games.steam_id))
print('users_items_raw.items = items_anexados_a_steam_games.items?', users_items_raw['items'].equals(items_anexados_a_steam_games['items']))


Las columnas items tienen valores nulos?

users_items_raw.info()

items_anexados_a_steam_games.info()

No poseen valores nulos pero podemos notar que difieren en el tipo de dato.

print("El tipo de la columna 'users_items.items_count' es ", users_items_raw.items_count.dtypes)
print("El tipo de la columna 'items_anexados_a_steam_games.items_count' es ", items_anexados_a_steam_games.items_count.dtypes)

Cambiaremos el tipo de 'users_items_raw.items_count' para hacer nuevamente la comparación.

users_items_raw_comparacion = users_items_raw
users_items_raw_comparacion.items_count = users_items_raw.items_count.astype("float64")
print("El tipo de la columna 'users_items_raw_comparacion.items_count' es ", users_items_raw.items_count.dtypes)

print('users_items.items_count = items_anexados_a_steam_games.items_count?', users_items_raw.items_count.equals(items_anexados_a_steam_games.items_count))

Hemos confirmado que la tabla items es la que se encuentra anexada a games por lo que procederemos a eliminarla.

----

steam_games_df = gamesRaw[~gamesRaw['user_id'].notna()] #  eliminamos las filas donde 'user_id' posee valores distintos de NaN
steam_games_df.drop(columns=['user_id', 'steam_id','items', 'items_count'], inplace=True)
steam_games_df = steam_games_df.rename(columns={'id':'item_id'})
steam_games_df.reset_index(drop=True, inplace=True)
steam_games_df.head(2)

steam_games_df.info()

(120444-32135) / 120444

Al eliminar las columnas 'user_id', 'steam_id', 'items' y 'items_count', hemos reducido un 75% del dataframe games.

---------

Analisis de los valores duplicados de steam_games_df

# No podemos usar el siguiente código porque una columna contiene listas
# (
#     steam_games_df
#     .duplicated()
#     .sum()
# )
# Asi que analizaremos los duplicados por columnas

(
    steam_games_df
    .duplicated(subset='item_id')
    .sum()
)


Existen items con el mismo item_id

steam_games_df[steam_games_df.duplicated(subset='item_id',keep=False)]

¿En qué se diferencian estas filas?

fila1 = steam_games_df.loc[13894]
fila2 = steam_games_df.loc[14573]

columnas_diferentes = fila1 != fila2 # me devuelve true/false 
columnas_diferentes = columnas_diferentes[columnas_diferentes] # filtra las columnas que son true

print("Columnas en las que difieren las filas:")
print(columnas_diferentes.index.tolist()) # Hace una lista con los nombres de las columnas diferentes

print (steam_games_df.at[13894,'discount_price'])
print (steam_games_df.at[14573,'discount_price'])

print (steam_games_df.at[13894,'url'])
print (steam_games_df.at[14573,'url'])

Hemos probado ambos links y llevan al mismo sitio por lo que solo dejaremos un registro

steam_games_df = steam_games_df.drop([74,14573])

Observando la tabla podemos deducir que el id del juego se encuentra en la url por lo que se completará con el id 200260 el valor faltante en la fila 30961

steam_games_df.at[30961,'item_id'] = 30961

steam_games_df = steam_games_df.reset_index(drop=True)


steam_games_df.head()

-------

Análisis de valores nulos de steam_games_df sin las columnas anexadas.

(
    steam_games_df
    .isnull()
    .sum()
)

steam_games_df[steam_games_df.item_id.isna()]

Eliminaremos este registro ya que la mayoría de sus valores son nulos.

steam_games_df = steam_games_df.dropna(subset='item_id')

Mas adelante, en la creación de tablas para los endpoint se tomaran decisiones pertinentes sobres los valores nulos que presenta este df en las restantes columnas.

steam_games_df.info()

steam_games_df.item_id = steam_games_df.item_id.astype("int")


------

Los Datasets users_items y raw_steam_games_df estan relacionados. Entendemos que si un usuario no esta registrado y no ha consumido ningun item no debería realizar comentarios del mismo. Por lo que revisaremos si en el df users_reviews existen usuarios que no se encuentran en users_items.

users_items.head(2)

user_reviews.head(2)

# users_de_users_items = list(users_items.user_id.drop_duplicates())
# users_de_user_reviews = list(user_reviews.user_id.drop_duplicates())
print (f' cantidad de users de users_items: ',len(users_de_users_items))
print (f' cantidad de users de user_reviews: ',len(users_de_user_reviews))
# Convierte las listas en conjuntos
users_de_user_reviews_set = set(users_de_user_reviews)
users_de_users_items_set = set(users_de_users_items)

# Encuentra los elementos en lista1 que no están en lista2
usuarios_faltantes = users_de_user_reviews_set - users_de_users_items_set

print("Cantidad de usuarios en reviews.user_id que no están en users_items.user_id:", len(usuarios_faltantes))

Efectivamente tenemos usuarios no registrados en users_items que se encuentran en user_reviews. Procederemos a eliminar estos usuarios para este primer mvp pero guardaremos los ids para informar de los 2845 usuarios que se encuentran en esta situación.

users_faltantes_en_users_df = user_reviews[(user_reviews.user_id.isin(usuarios_faltantes))].reset_index(drop=True)

users_faltantes_en_users_df

user_reviews = user_reviews[~(user_reviews.user_id.isin(usuarios_faltantes))]
user_reviews

Guardamos los df en archivos csv

users_faltantes_en_users_df.to_csv("data/users_faltantes_en_users_df.csv",index=False)

user_reviews.to_csv("data/raw_user_reviews.csv",index=False)
steam_games_df.to_csv("data/raw_steam_games.csv",index=False)
users_items.to_csv("data/raw_user_items.csv",index=False)
