==> 3_relaciones_entre_tablas.ipynb <==
# Librerías y configuraciones


import pandas as pd
import warnings

pd.set_option('display.max_colwidth', 100)
warnings.filterwarnings('ignore')

---

# Carga de datos

df_genres = pd.read_csv('data/df_genres.csv')
df_item_genre = pd.read_csv('data/df_item_genre.csv')
df_items = pd.read_csv('data/df_items.csv')
df_reviews = pd.read_csv('data/df_reviews.csv')
df_users_items = pd.read_csv('data/df_users_items.csv')
df_users = pd.read_csv('data/df_users.csv')

----

Funciones auxiliares

def items_faltantes(columna_principal,columna_relacionada):
    list_elementos_unicos_de_columna_principal = list(columna_principal.unique())
    list_elementos_unicos_de_columna_relacionada = list(columna_relacionada.unique())
    print(len(list_elementos_unicos_de_columna_principal))
    print(len(list_elementos_unicos_de_columna_relacionada))

    # Convertir las listas a conjuntos
    list_elementos_unicos_de_columna_principal = set(list_elementos_unicos_de_columna_principal)
    list_elementos_unicos_de_columna_relacionada = set(list_elementos_unicos_de_columna_relacionada)

    # Encontrar los elementos que están en set_items_en_df_item_genre pero no en set_items_en_df_items
    elementos_no_en_df_items =   list_elementos_unicos_de_columna_relacionada - list_elementos_unicos_de_columna_principal
    # Obtener la cantidad y los elementos que no están en df_items
    cantidad_elementos_no_en_df_items = len(elementos_no_en_df_items)
    lista_elementos_no_en_df_items = list(elementos_no_en_df_items)

    print("Cantidad de items de la columna relacionada que no se encuentran en la columna principal ", (cantidad_elementos_no_en_df_items))
    print("Elementos que faltan en la columna principal:", lista_elementos_no_en_df_items)
    return lista_elementos_no_en_df_items

----

Las tablas:
- df_genres
- df_item_genre
- df_items
- df_reviews 
- df_users_items
- df_users 

estan relacionadas por lo que no deberíamos tener:
- df_items
    - items en df_item_genre que no esten en df_items
    - items en df_users_items que no esten en df_items
    - items en df_reviews que no esten en  df_items
- df_users
    - users en reviews que no esten en users
    - users en users_items que no esten en users

    - items en df_item_genre que no esten en df_items


items_de_df_item_genre_que_no_estan_en_df_items = items_faltantes(df_items.item_id,df_item_genre.item_id)

    - items en df_users_items que no esten en df_items


items_de_df_users_items_que_no_estan_en_df_items = items_faltantes(df_items.item_id,df_users_items.item_id)

    - items en df_reviews que no esten en  df_items


items_de_df_reviews_que_no_estan_en_df_items = items_faltantes(df_items.item_id,df_reviews.item_id)

crearemos una lista de estos items

items_sobrantes = set(items_de_df_item_genre_que_no_estan_en_df_items + items_de_df_users_items_que_no_estan_en_df_items +items_de_df_reviews_que_no_estan_en_df_items)

len(items_sobrantes)

df_items[df_items.item_id==245760]

Existen 2867 items que se encuentran en las tablas relacionadas pero no en la tabla donde se almacena la info de items.
Para que la información sea consistente eliminaremos estos items de los df.

Eliminamos items de df_item_genre

df_item_genre[df_item_genre.item_id.isin(items_sobrantes)]

df_item_genre = df_item_genre[~(df_item_genre.item_id.isin(items_sobrantes))]

Eliminamos items de df_users_item

df_users_items[df_users_items.item_id.isin(items_sobrantes)]

df_users_items = df_users_items[~(df_users_items.item_id.isin(items_sobrantes))]

Eliminamos items de df_reviews

df_reviews[df_reviews.item_id.isin(items_sobrantes)]

df_reviews = df_reviews[~(df_reviews.item_id.isin(items_sobrantes))]

    - users en reviews que no esten en users


items_de_df_reviews_que_no_estan_en_df_users = items_faltantes(df_users.user_id,df_reviews.user_id)

    - users en users_items que no esten en users

items_de_df_users_items_que_no_estan_en_df_users = items_faltantes(df_users.user_id,df_users_items.user_id)

Podemos ver que no existen usuarios en df_reviews y df_users_items que no se encuentren en df_users

Las tablas:
- df_genres
- df_item_genre
- df_items
- df_reviews 
- df_users_items
- df_users 

df_genres.to_csv("data/df_genres.csv",index=False)
df_item_genre.to_csv("data/df_item_genre.csv",index=False)
df_users_items.to_csv("data/df_users_items.csv",index=False)
df_reviews.to_csv("data/df_reviews.csv",index=False)
df_users_items.to_csv("data/df_users_items.csv",index=False)
